+++ src/main/java/com/minhcraft/beyondbetatweaks/BeyondBetaTweaks.java
@@ -3,9 +3,9 @@ package com.minhcraft.beyondbetatweaks;
 import com.minhcraft.beyondbetatweaks.config.ModConfig;
 import com.minhcraft.beyondbetatweaks.config.RecipeCustomSortingConfigLoader;
 import com.minhcraft.beyondbetatweaks.network.ModNetworking;
-import com.minhcraft.beyondbetatweaks.register.ModItems;
-import com.minhcraft.beyondbetatweaks.register.ModRegistry;
-import com.minhcraft.beyondbetatweaks.register.ModSounds;
+import com.minhcraft.beyondbetatweaks.register.BeyondBetaItems;
+import com.minhcraft.beyondbetatweaks.register.BeyondBetaRegistry;
+import com.minhcraft.beyondbetatweaks.register.BeyondBetaSounds;
 import com.minhcraft.beyondbetatweaks.world.BeyondBetaCarvers;
 import eu.midnightdust.lib.config.MidnightConfig;
 import net.fabricmc.api.ModInitializer;
@@ -30,9 +30,9 @@ public class BeyondBetaTweaks implements ModInitializer {
 		// Proceed with mild caution.
 
 		LOGGER.info("Initializing Beyond Beta Tweaks");
-		ModSounds.init();
-		ModItems.init();
-		ModRegistry.init();
+		BeyondBetaSounds.init();
+		BeyondBetaItems.init();
+		BeyondBetaRegistry.init();
 		BeyondBetaCarvers.register();
 
 		ModNetworking.registerServerReceivers();